chatbot/src/requirements.txt
google-cloud-dialogflow
nltk
spacy
requests
rich
